Refactor history and add replace doc command
